Bug-fix and much tighter checks on sequelae probability reading from XML for old case-management model. This will change output "nSeq" due to bug-fix and may cause scenarios to be rejected.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@712 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/Clinical/OldCaseManagement.h
@@ -137,7 +137,7 @@ private:
   
   /// Age bounds of probSequelae* parameters
   //@{
-  static const int NUM_SEQUELAE_AGE_GROUPS = 2;
+  static const size_t NUM_SEQUELAE_AGE_GROUPS = 2;
   static const int SEQUELAE_AGE_BOUND[NUM_SEQUELAE_AGE_GROUPS];
   //@}
